chore: added more logging to the api
